Refactor: extract `TokenTooltip` from `TokensTab`
